+++ docs/advanced_topics/amending_static_inputs/plan.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python
+from stepup.core.api import static, amend
+
+static("config.txt")
+amend(inp="config.txt")
+with open("config.txt") as fh:
+    print(fh.read().strip())