+++ tests/cases/error_cyclic_early/plan.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-from stepup.core.api import static, amend
-
-static("README.md")
-amend(inp="README.md")